moved PhotosDiv to components
