+++ src/pages/Home.js
@@ -1,12 +1,12 @@
 import { Routes, Route } from 'react-router-dom'
 import Login from './Login'
 import MapBoard from '../components/MapBoard'
-import PhotosDiv from './PhotosDiv'
+import PhotosDiv from '../components/PhotosDiv'
 
 const Home = () => {
 
     return (
-        <div>
+        <div className='homePage'>
             <Routes>
                 <Route path='/register/login' element={
                     <Login />
